requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy==1.24.4
Pillow==10.2.0
scikit-learn==1.3.2
matplotlib==3.8.2
gymnasium==0.29.1
stable-baselines3==2.2.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain==0.1.4
langchain-community==0.0.16
faiss-cpu==1.7.4
sentence-transformers==2.3.1
requests==2.31.0
rasterio==1.3.9
gdown==5.1.0
